+++ dev/Neuron.js
@@ -15,7 +15,10 @@ class Neuron {
             this.weights = [...new Array(size)].map(v => Math.random()*0.2-0.1)
             this.bias = Math.random()*0.2-0.1
         }
+
         this.deltaWeights = this.weights.map(v => 0)
+        this.weightGains = [...new Array(size)].map(v => 1)
+        this.biasGain = 1
     }
 }