+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Upcoming
 ---
+#### Network
+- Added adaptiveLR as a network configuration
+- Added gain as an adaptive learning rate function
 #### NetMath
-- Added Mean Squared Error
+- Added Mean Squared Error cost function
 
 # v1.0.0
 ----